refactor: Cars 초기화 CarService로 이동
