+++ src/main/java/service/CarService.java
@@ -2,17 +2,42 @@ package service;
 
 import dto.CarDto;
 import dto.WinnerCarDto;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
+import exception.DuplicateCarNameException;
 import model.Car;
 import model.Cars;
 import utils.RacingNumberGenerator;
 
 public class CarService {
 
+    private static final String SEPARATOR = ",";
+
     private Cars cars;
 
-    public void generateCars(List<Car> car) {
-        this.cars = new Cars(car);
+    public void initCars(String inputCarsName) {
+        String[] carsName = inputCarsName.split(SEPARATOR);
+        checkDuplication(carsName);
+        this.cars = new Cars(mapToCars(carsName));
+    }
+
+    private List<Car> mapToCars(String[] carsName) {
+        return Arrays.stream(carsName)
+                .map(Car::new)
+                .collect(Collectors.toList());
+    }
+
+    private void checkDuplication(String[] carsName) {
+        if (getDistinctCarsCount(carsName) != carsName.length) {
+            throw new DuplicateCarNameException();
+        }
+    }
+
+    private long getDistinctCarsCount(String[] carsName) {
+        return Arrays.stream(carsName)
+                .distinct()
+                .count();
     }
 
     public void race(RacingNumberGenerator generator) {
