+++ src/main/java/view/InputView.java
@@ -8,7 +8,6 @@ public class InputView {
             "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
     private static final String INPUT_ROUND_MESSAGE =
             "시도할 회수는 몇회인가요?";
-
     private static final Scanner scanner = new Scanner(System.in);
 
     public static String inputCarsName() {
